pyproject.toml
[project]
name = "test-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fschat[model-worker,webui]>=0.2.36",
    "google-generativeai>=0.8.3",
    "natsort>=8.4.0",
    "openai>=1.59.7",
    "pylcs>=0.1.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.3.1",
    "torch>=2.5.1",
    "transformers>=4.48.0",
]
